Refactor podcast external links functionality & upgrade to typescript
