+++ src/app/routes/onDemandAudio/podcastExternalLinks/arabic.ts
@@ -1,0 +1,133 @@
+const externalLinks = {
+  p02pc9qc: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/3k7pugwtP9ZouNIj2x97I6',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl: 'https://podcasts.apple.com/us/podcast/bbc-xtra/id262251878',
+      linkType: 'apple',
+    },
+  ],
+  p08z9x26: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/3X9hcuWajnQ1QytkmvmV5k',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/ae/podcast/%D8%AE%D8%B1%D8%A7%D9%81%D8%A7%D8%AA/id1541601477',
+      linkType: 'apple',
+    },
+  ],
+  p086jpqy: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/1Ipw9UySbhZvZgEBAPKnYK',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D8%A7%D9%84%D9%86%D8%B4%D8%B1%D8%A9-%D8%A7%D9%84%D9%8A%D9%88%D9%85%D9%8A%D8%A9-%D9%84%D9%81%D9%8A%D8%B1%D9%88%D8%B3-%D9%83%D9%88%D8%B1%D9%88%D9%86%D8%A7/id1502996667',
+      linkType: 'apple',
+    },
+  ],
+  p08gp8wx: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/2Hd2doehQOTkRsRni0rICY',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/%D8%AE%D8%B1%D8%A7%D9%81%D8%A7%D8%AA-%D9%83%D9%88%D8%B1%D9%88%D9%86%D8%A7/id1518556307',
+      linkType: 'apple',
+    },
+  ],
+  p09cgw89: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/3E1HIl16xfaFG6RXImxpVI',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/%D8%A3%D8%A8%D8%B9%D8%A7%D8%AF-%D8%A7%D9%84%D8%B5%D9%88%D8%B1%D8%A9/id1561146726',
+      linkType: 'apple',
+    },
+  ],
+  p09m6x31: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/3VsAmuPq7qHIRofC5biK8c',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D8%A8%D9%88%D8%AF%D9%83%D8%A7%D8%B3%D8%AA-%D8%B9-%D9%84%D8%A7%D9%82%D8%A7%D8%AA/id1573549106',
+      linkType: 'apple',
+    },
+  ],
+  p09w8yvk: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/0dyrgq9nQMdZ5b7Hfcz6q6',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D8%A5%D9%83%D8%B3%D8%AA%D8%B1%D8%A7-%D9%84%D9%8A%D9%81%D9%84%D8%B2-xtra-levels/id1586351356',
+      linkType: 'apple',
+    },
+  ],
+  p0b3xdrj: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4Hftvi6T6kjQRogRL1yBVx',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/%D9%85%D8%B1%D8%A7%D9%87%D9%82%D8%AA%D9%8A/id1595160361',
+      linkType: 'apple',
+    },
+  ],
+  p0c9wp0l: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/3SPVfsKTZ1V6DPvIrvVyAb',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/تغيير-بسيط-a-simple-change/id1626812363',
+      linkType: 'apple',
+    },
+  ],
+  p0h6d6nm: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/2IidLqm28ZVS9E75ZJ9Fqj',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/%D9%8A%D8%B3%D8%AA%D8%AD%D9%82-%D8%A7%D9%84%D8%A7%D9%86%D8%AA%D8%A8%D8%A7%D9%87/id1726949222',
+      linkType: 'apple',
+    },
+  ],
+};
+
+export default externalLinks;
