+++ src/app/routes/onDemandAudio/podcastExternalLinks/burmese.ts
@@ -1,0 +1,30 @@
+const externalLinks = {
+  p02pc9lh: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4SMim5Kg9ex8VwO3D5Kc8n',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/my/podcast/burmese-evening-broadcast/id352604027',
+      linkType: 'apple',
+    },
+  ],
+  p02p9f6q: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/6okAEFPPbfms2Ik4ZoAD3t',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/burmese-morning-broadcast/id352604028',
+      linkType: 'apple',
+    },
+  ],
+};
+
+export default externalLinks;
