+++ src/app/routes/onDemandAudio/podcastExternalLinks/gahuza.ts
@@ -1,0 +1,68 @@
+const externalLinks = {
+  p07yh8hb: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/3X2EmcD5e9YZjbLgX9SZls',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/baza-muganga/id1492445243',
+      linkType: 'apple',
+    },
+  ],
+  p02pcb5c: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4WIzS3ZAXxw1fxUx4FCTBY',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl: 'https://podcasts.apple.com/gb/podcast/imvo-nimvano/id981906761',
+      linkType: 'apple',
+    },
+  ],
+  p07yjlmf: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/5cW7niDjkEF35hGC4rOUlD',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/ikiganiro-cyabagore/id1492486017',
+      linkType: 'apple',
+    },
+  ],
+  p07yhgwh: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/1HWMG5MGbcsGOVKkHeTiof',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/amakuru-kuri-bbc-gahuzamiryango/id1492454113',
+      linkType: 'apple',
+    },
+  ],
+  p07yjfjq: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/5RVTKebCVwCBvC0zHnCKfz',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/ikinamico-urunana/id1492479210',
+      linkType: 'apple',
+    },
+  ],
+};
+
+export default externalLinks;
