+++ src/app/routes/onDemandAudio/podcastExternalLinks/hausa.ts
@@ -1,0 +1,17 @@
+const externalLinks = {
+  p08mlgcb: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/1JGid3GVGVdab0FycTUY4L',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/korona-ina-mafita/id1526116287',
+      linkType: 'apple',
+    },
+  ],
+};
+
+export default externalLinks;
