+++ src/app/routes/onDemandAudio/podcastExternalLinks/index.test.ts
@@ -4,13 +4,14 @@ import zhongwenExternalLinks from './zhongwen';
 
 describe('getPodcastExternalLinks', () => {
   it('should return the links', async () => {
-    const links = await getPodcastExternalLinks(
-      'hausa',
-      'p08mlgcb',
-      'p0967t2j',
-    );
+    const links = await getPodcastExternalLinks({
+      service: 'hausa',
+      brandId: 'p08mlgcb',
+      versionId: 'p0967t2j',
+    });
+
     const expectedLinks = [
-      ...hausaExternalLinks.default.p08mlgcb,
+      ...hausaExternalLinks.p08mlgcb,
       {
         linkText: 'RSS',
         linkUrl: 'https://podcasts.files.bbci.co.uk/p08mlgcb.rss',
@@ -27,17 +28,29 @@ describe('getPodcastExternalLinks', () => {
   });
 
   it('should return an empty array when service is incorrect', async () => {
-    const links = await getPodcastExternalLinks('foo', 'p08mlgcb');
+    const links = await getPodcastExternalLinks({
+      // @ts-expect-error partial data for testing purposes
+      service: 'foo',
+      brandId: 'p08mlgcb',
+    });
     expect(links).toEqual([]);
   });
 
   it('should return an empty array when brand is null', async () => {
-    const links = await getPodcastExternalLinks('hausa', null);
+    const links = await getPodcastExternalLinks({
+      service: 'hausa',
+      // @ts-expect-error partial data for testing purposes
+      brandId: null,
+    });
     expect(links).toEqual([]);
   });
 
   it('should return rss feed when brand is not found', async () => {
-    const otherLinks = await getPodcastExternalLinks('hausa', 'bar', 'h455a');
+    const otherLinks = await getPodcastExternalLinks({
+      service: 'hausa',
+      brandId: 'bar',
+      versionId: 'h455a',
+    });
     expect(otherLinks).toEqual([
       {
         linkText: 'RSS',
@@ -54,12 +67,12 @@ describe('getPodcastExternalLinks', () => {
   });
 
   it('should return the links with variants', async () => {
-    const links = await getPodcastExternalLinks(
-      'zhongwen',
-      'p02pc9xp',
-      'p0967t2j',
-      'simp',
-    );
+    const links = await getPodcastExternalLinks({
+      service: 'zhongwen',
+      variant: 'simp',
+      brandId: 'p02pc9xp',
+      versionId: 'p0967t2j',
+    });
     const expectedLinks = [
       ...zhongwenExternalLinks.simp.p02pc9xp,
       {
@@ -78,12 +91,11 @@ describe('getPodcastExternalLinks', () => {
   });
 
   it('should return the correct download link', async () => {
-    const links = await getPodcastExternalLinks(
-      'burmese',
-      'p02pc9lh',
-      'p0967t2j',
-      undefined,
-    );
+    const links = await getPodcastExternalLinks({
+      service: 'burmese',
+      brandId: 'p02pc9lh',
+      versionId: 'p0967t2j',
+    });
     expect(links[3].linkUrl).toEqual(
       'https://open.live.bbc.co.uk/mediaselector/6/redir/version/2.0/mediaset/audio-nondrm-download-low/proto/https/vpid/p0967t2j.mp3',
     );
