+++ src/app/routes/onDemandAudio/podcastExternalLinks/kyrgyz.ts
@@ -1,0 +1,30 @@
+const externalLinks = {
+  p09hxksq: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/0Ns5LMWc959Cummte0k9Q1',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/25-%D0%B6%D1%8B%D0%BB-%D0%BE%D0%B1%D0%BE%D0%B4%D0%BE/id1568103058',
+      linkType: 'apple',
+    },
+  ],
+  p0c80v81: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/7nYFDxTuJnmWtS5lhxtjmB',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/би-би-си-дүрбүсүндө/id1625319429',
+      linkType: 'apple',
+    },
+  ],
+};
+
+export default externalLinks;
