+++ src/app/routes/onDemandAudio/podcastExternalLinks/marathi.ts
@@ -1,0 +1,74 @@
+const externalLinks = {
+  p09431p4: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4f0yoffmGYYqaTAwI4Ngal',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%E0%A4%A4-%E0%A4%A8-%E0%A4%97-%E0%A4%B7-%E0%A4%9F/id1550093946',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Gaana',
+      linkUrl: 'https://gaana.com/podcast/-season-1-2021-17',
+      linkType: 'gaana',
+    },
+    {
+      linkText: 'JioSaavn',
+      linkUrl: 'https://www.jiosaavn.com/shows/तीन-गोष्टी/1/5k2iS6,xzZc_',
+      linkType: 'jiosaavn',
+    },
+  ],
+  p09437hm: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/0LK47dXlIqm0Gm1DLeYAIS',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%E0%A4%B8-%E0%A4%AA-%E0%A4%97-%E0%A4%B7-%E0%A4%9F/id1550094137',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Gaana',
+      linkUrl: 'https://gaana.com/podcast/-season-1-2021-18',
+      linkType: 'gaana',
+    },
+    {
+      linkText: 'JioSaavn',
+      linkUrl: 'https://www.jiosaavn.com/shows/सोपी-गोष्ट/1/kVtGnMWt4kE_',
+      linkType: 'jiosaavn',
+    },
+  ],
+  p0b1s4nm: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/2HlWdNCZ3DpwUMXL4FlHs6',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%E0%A4%97-%E0%A4%B7-%E0%A4%9F-%E0%A4%A6-%E0%A4%A8-%E0%A4%AF-%E0%A4%9A/id1595097031',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Gaana',
+      linkUrl: 'https://gaana.com/podcast/-season-1-2021-271',
+      linkType: 'gaana',
+    },
+    {
+      linkText: 'JioSaavn',
+      linkUrl:
+        'https://www.jiosaavn.com/shows/%E0%A4%97%E0%A5%8B%E0%A4%B7%E0%A5%8D%E0%A4%9F-%E0%A4%A6%E0%A5%81%E0%A4%A8%E0%A4%BF%E0%A4%AF%E0%A5%87%E0%A4%9A%E0%A5%80/1/NeVi8wudirQ_',
+      linkType: 'jiosaavn',
+    },
+  ],
+};
+
+export default externalLinks;
