+++ src/app/routes/onDemandAudio/podcastExternalLinks/persian.ts
@@ -1,0 +1,111 @@
+const externalLinks = {
+  p02pc9wf: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/2z6NRjAYvSaInglrZOfawC',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl: 'https://podcasts.apple.com/us/podcast/pargar/id634850665',
+      linkType: 'apple',
+    },
+  ],
+  p036kbdd: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/1y5MuXLzLPJsw1YbtPNMe6',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/in/podcast/be-ebarat-e-digar/id1055378813',
+      linkType: 'apple',
+    },
+  ],
+  p05gyy09: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/08dvs0rLyyNni2I3lhteen',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D8%B5%D9%81%D8%AD%D9%87-%DB%B2-page-2/id1288618875',
+      linkType: 'apple',
+    },
+  ],
+  p0703hz7: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/5oMeCbKkRkqx8X7IKpAZzB',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D8%AF%D8%A7%D8%B3%D8%AA%D8%A7%D9%86-%D8%A7%D9%86%D9%82%D9%84%D8%A7%D8%A8-%D8%A8%D9%87-%D8%B1%D9%88%D8%A7%DB%8C%D8%AA-%D8%A8%DB%8C-%D8%A8%DB%8C-%D8%B3%DB%8C/id1451707438',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Castbox',
+      linkUrl:
+        'https://castbox.fm/channel/%D9%85%D8%B3%D8%AA%D9%86%D8%AF-id2007735?country=gb',
+      linkType: 'castbox',
+    },
+  ],
+  p02pc9mc: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/0cHK9mp72ixQv8PVd7Bb3N',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/bbc-persian-radio/id262741509',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Castbox',
+      linkUrl:
+        'https://castbox.fm/channel/رادیو-فارسی-بیبیسی-id574405?utm_source=website&utm_medium=dlink&utm_campaign=web_share&utm_content=%D8%B1%D8%A7%D8%AF%DB%8C%D9%88%20%D9%81%D8%A7%D8%B1%D8%B3%DB%8C%20%D8%A8%DB%8C%E2%80%8C%D8%A8%DB%8C%E2%80%8C%D8%B3%DB%8C-CastBox_FM',
+      linkType: 'castbox',
+    },
+  ],
+  p09v5lgp: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4k2EWdvmGiWLgXpUrm8VMS',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D8%A8%DB%8C%D9%86-%D8%B3%D8%B7%D9%88%D8%B1/id1584665242',
+      linkType: 'apple',
+    },
+  ],
+  p0bw80rj: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4kInUtgdsMnUxlA7sBKxLL',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl: 'https://podcasts.apple.com/us/podcast/بین-سطور/id1615070108',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Castbox',
+      linkUrl:
+        'https://castbox.fm/channel/%D8%B4%DB%8C%D8%B1%D8%A7%D8%B2%D9%87-id4841117?country=us',
+      linkType: 'castbox',
+    },
+  ],
+};
+
+export default externalLinks;
