+++ src/app/routes/onDemandAudio/podcastExternalLinks/portuguese.ts
@@ -1,0 +1,90 @@
+const externalLinks = {
+  p07r3r3t: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4KqTMEWcOUzHFHAPCt54Uk',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/br/podcast/que-hist%C3%B3ria/id1492686435',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'YouTube',
+      linkUrl:
+        'https://www.youtube.com/playlist?list=PLCX5XjxKTpTmh19ipbvr6OACmBWczOLJO',
+      linkType: 'youtube',
+    },
+    {
+      linkText: 'Amazon Music',
+      linkUrl:
+        'https://music.amazon.co.uk/podcasts/51c33ac0-d208-4026-bc4d-396205703a8f/que-hist%C3%B3ria',
+      linkType: 'amazon',
+    },
+  ],
+  p09qw1cn: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/54YV8Rd6zmvq9Ry55q9HQw',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl: 'https://podcasts.apple.com/us/podcast/bbc-l%C3%AA/id1586358051',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Amazon Music',
+      linkUrl:
+        'https://music.amazon.co.uk/podcasts/5082d646-cd10-4192-8f83-4812d02b6f83/bbc-l%C3%AA',
+      linkType: 'amazon',
+    },
+    {
+      linkText: 'Arquivo completo',
+      linkUrl: 'https://www.bbc.com/portuguese/topics/cxndrr1qgllt',
+      linkType: 'arquivocompleto',
+    },
+  ],
+  p0cx5pt6: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/5OZrG0afS38YHeaCnBQbBz',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/as-estranhas-origens-das-guerras-culturais/id1642483075',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Amazon Music',
+      linkUrl:
+        'https://music.amazon.co.uk/podcasts/b4b5f9ad-7899-43d7-8ffb-169a4d93db62/as-estranhas-origens-das-guerras-culturais',
+      linkType: 'amazon',
+    },
+  ],
+  p0cyhvny: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4909a8iZAdJ01DLyJxiwcd',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/us/podcast/brasil-partido/id1643515422',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Amazon Music',
+      linkUrl:
+        'https://music.amazon.co.uk/podcasts/49d38ff1-5646-4fdc-a9df-bdb5ccffda34/brasil-partido',
+      linkType: 'amazon',
+    },
+  ],
+};
+
+export default externalLinks;
