+++ src/app/routes/onDemandAudio/podcastExternalLinks/types.ts
@@ -1,0 +1,16 @@
+import { Services, Variants } from '#app/models/types/global';
+
+export type PodcastExternalLinksParams = {
+  service: Services;
+  variant?: Variants;
+  brandId: string;
+  versionId: string;
+};
+
+export type ExternalLinks = {
+  [key: string]: {
+    linkText: string;
+    linkUrl: string;
+    linkType: string;
+  };
+};
