+++ src/app/routes/onDemandAudio/podcastExternalLinks/ukrainian.ts
@@ -1,0 +1,17 @@
+const externalLinks = {
+  p09jsy3h: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/3q7550hSqicRIJSFsQ5BVr',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D1%89%D0%BE-%D1%86%D0%B5-%D0%B1%D1%83%D0%BB%D0%BE/id1569801762',
+      linkType: 'apple',
+    },
+  ],
+};
+
+export default externalLinks;
