+++ src/app/routes/onDemandAudio/podcastExternalLinks/urdu.ts
@@ -1,0 +1,40 @@
+const externalLinks = {
+  p0c04t7w: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/4izrTfBmEKZHYCAUiuc4gw',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%DA%88%D8%B1%D8%A7%D9%85%DB%81-%DA%A9%D9%88%D8%A6%DB%8C%D9%86-drama-queen/id1618467017',
+      linkType: 'apple',
+    },
+    {
+      linkText: 'Patari',
+      linkUrl: 'https://patari.pk/home/song/625a05eb4950df5b58b97a4a',
+      linkType: 'patari',
+    },
+    {
+      linkText: 'اپنے دل کی بات ہمارے ساتھ شیئر کیجیے',
+      linkUrl: 'https://www.bbc.com/urdu/send/u106609198',
+      linkType: 'form',
+    },
+  ],
+  p0clc9zw: [
+    {
+      linkText: 'Spotify',
+      linkUrl: 'https://open.spotify.com/show/7iOhf64Hu8wBCZDQSm0Dod',
+      linkType: 'spotify',
+    },
+    {
+      linkText: 'Apple',
+      linkUrl:
+        'https://podcasts.apple.com/gb/podcast/%D8%A8%D8%A7%D8%AA-%D8%B3%D8%B1%D8%AD%D8%AF-%D9%BE%D8%A7%D8%B1/id1634186835',
+      linkType: 'apple',
+    },
+  ],
+};
+
+export default externalLinks;
