--- src/app/routes/onDemandAudio/tempData/podcastExternalLinks/hindi.js
@@ -1,263 +1,0 @@
-const externalLinks = {
-  default: {
-    p0552909: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/1eqTGd8QFThO8jOOl7zWoY',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/%E0%A4%B5-%E0%A4%B5-%E0%A4%9A%E0%A4%A8/id1245120783',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/%E0%A4%B5%E0%A4%BF%E0%A4%B5%E0%A5%87%E0%A4%9A%E0%A4%A8%E0%A4%BE/1/pkViS6nFvD0_',
-        linkType: 'jiosaavn',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/podcast/-season-1-2019-2019',
-        linkType: 'gaana',
-      },
-    ],
-    p055260j: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/0SDaLngHehBulm5LXpyqSV',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/%E0%A4%AC-%E0%A4%AC-%E0%A4%B8-%E0%A4%8F%E0%A4%95-%E0%A4%AE-%E0%A4%B2-%E0%A4%95-%E0%A4%A4/id1244954195',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/%E0%A4%AC%E0%A5%80%E0%A4%AC%E0%A5%80%E0%A4%B8%E0%A5%80-%E0%A4%8F%E0%A4%95-%E0%A4%AE%E0%A5%81%E0%A4%B2%E0%A4%BE%E0%A4%95%E0%A4%BE%E0%A4%A4/1/9x-ipVkS1f8_',
-        linkType: 'jiosaavn',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/podcast/-season-1-2020',
-        linkType: 'gaana',
-      },
-    ],
-    p05528hs: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/1izS1hnTt4hDVwysWQ8HYi',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%E0%A4%AC-%E0%A4%AC-%E0%A4%B8-%E0%A4%A8-%E0%A4%AF-%E0%A4%9C-%E0%A4%AE-%E0%A4%95%E0%A4%B0-%E0%A4%B8/id1245120209',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/%E0%A4%AC%E0%A5%80%E0%A4%AC%E0%A5%80%E0%A4%B8%E0%A5%80-%E0%A4%A8%E0%A5%8D%E0%A4%AF%E0%A5%82%E0%A4%9C%E0%A4%BC-%E0%A4%AE%E0%A5%87%E0%A4%95%E0%A4%B0%E0%A5%8D%E0%A4%B8/1/d8gj1oPueo4_',
-        linkType: 'jiosaavn',
-      },
-    ],
-    p05523zq: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/4NwNiuICPqNRfGsg4Qm9wH',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%E0%A4%AC-%E0%A4%AC-%E0%A4%B8-70-%E0%A4%8F%E0%A4%AE%E0%A4%8F%E0%A4%AE/id1244939385',
-        linkType: 'apple',
-      },
-    ],
-    p05525mc: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/4ZERQGr2e0byGAtbTa76dR',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%E0%A4%A6-%E0%A4%A8-%E0%A4%AF-%E0%A4%9C%E0%A4%B9-%E0%A4%A8/id1244954172',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/%E0%A4%A6%E0%A5%81%E0%A4%A8%E0%A4%BF%E0%A4%AF%E0%A4%BE-%E0%A4%9C%E0%A4%B9%E0%A4%BE%E0%A4%A8/1/,2RtKVyy0fI_',
-        linkType: 'jiosaavn',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/podcast/-season-1-2020-3',
-        linkType: 'gaana',
-      },
-    ],
-    p08s9wv2: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/2J22D8PZYQdvx4zICcRGon',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/se/podcast/%E0%A4%95%E0%A4%B9-%E0%A4%A8-%E0%A4%9C-%E0%A4%A6%E0%A4%97-%E0%A4%95/id1533103067',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/%E0%A4%95%E0%A4%B9%E0%A4%BE%E0%A4%A8%E0%A5%80-%E0%A4%9C%E0%A4%BC%E0%A4%BF%E0%A4%82%E0%A4%A6%E0%A4%97%E0%A5%80-%E0%A4%95%E0%A5%80/1/UTAeq5YqNqw_',
-        linkType: 'jiosaavn',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/podcast/-season-1-2021-4',
-        linkType: 'gaana',
-      },
-    ],
-    p09ds7zx: [
-      {
-        linkText: 'Spotify',
-        linkUrl:
-          'https://open.spotify.com/show/6BGVmN2D82U8hGgDBA4N07?si=9_IClv8KTQyFS4N9jQiSpw',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%E0%A4%A6-%E0%A4%A8%E0%A4%AD%E0%A4%B0-%E0%A4%AA-%E0%A4%B0-%E0%A4%A6-%E0%A4%A8-%E0%A4%AA-%E0%A4%B0-%E0%A4%96-%E0%A4%AC%E0%A4%B0-dinbhar/id1563251973',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/%E0%A4%A6%E0%A4%BF%E0%A4%A8-%E0%A4%AD%E0%A4%B0/1/ZBludb0NoRM_',
-        linkType: 'jiosaavn',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/podcast/-season-1-2019',
-        linkType: 'gaana',
-      },
-    ],
-
-    p0c0530h: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/02iGEmdv3do4InPguSE1cs',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%E0%A4%A1-%E0%A4%B0-%E0%A4%AE-%E0%A4%95-%E0%A4%B5-%E0%A4%A8-drama-queen/id1618467926',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'अपने दिल की बात हमसे शेयर करें',
-        linkUrl: 'https://www.bbc.com/hindi/send/u107107375',
-        linkType: 'form',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/season/drama-queen-season-1-2022',
-        linkType: 'gaana',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.saavn.com/s/show/ड्रामा-क्वीन-drama-queen/1/dZOptt08D4A_',
-        linkType: 'jiosaavn',
-      },
-    ],
-    p0clc83k: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/0sgucaqSDSOg30tyDdq960',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%E0%A4%AC-%E0%A4%A4-%E0%A4%B8%E0%A4%B0%E0%A4%B9%E0%A4%A6-%E0%A4%AA-%E0%A4%B0/id1634185584',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/%E0%A4%AC%E0%A4%BE%E0%A4%A4-%E0%A4%B8%E0%A4%B0%E0%A4%B9%E0%A4%A6-%E0%A4%AA%E0%A4%BE%E0%A4%B0/1/BzW104mU9GQ_',
-        linkType: 'jiosaavn',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/podcast/-season-1-2022-159',
-        linkType: 'gaana',
-      },
-    ],
-    p0frdsxv: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/0ngLvRDYAqCCp9j6Frs0uo',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/%E0%A4%95%E0%A4%B2-%E0%A4%95-%E0%A4%95-%E0%A4%B8-%E0%A4%A5-%E0%A4%AE-%E0%A4%88-%E0%A4%87-%E0%A4%A1-%E0%A4%AF%E0%A4%A8-%E0%A4%B2-%E0%A4%87%E0%A4%AB/id1690364464',
-        linkType: 'apple',
-      },
-    ],
-    p0fmrg25: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/5SaqjHgniLJIlPKGum79ia',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/%E0%A4%9B-%E0%A4%9F-%E0%A4%89%E0%A4%AE-%E0%A4%B0-%E0%A4%AC%E0%A4%A1-%E0%A4%9C-%E0%A4%A6%E0%A4%97/id1687504309',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Gaana',
-        linkUrl: 'https://gaana.com/podcast/-season-1-2023-153',
-        linkType: 'gaana',
-      },
-    ],
-    p0kjf03y: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/33zf2Pgpq1w0BXp1Vhe9vz',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/the-lens-मुकेश-शर्मा-के-साथ/id1791099679',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Jio Saavn',
-        linkUrl:
-          'https://www.jiosaavn.com/shows/the-lens-%e0%a4%ae%e0%a5%81%e0%a4%95%e0%a5%87%e0%a4%b6-%e0%a4%b6%e0%a4%b0%e0%a5%8d%e0%a4%ae%e0%a4%be-%e0%a4%95%e0%a5%87-%e0%a4%b8%e0%a4%be%e0%a4%a5/1/ESz7PpxJiiA_',
-        linkType: 'gaana',
-      },
-    ],
-  },
-};
-
-export default externalLinks;
