--- src/app/routes/onDemandAudio/tempData/podcastExternalLinks/indonesia.js
@@ -1,58 +1,0 @@
-const externalLinks = {
-  default: {
-    p02pc9v6: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/1llyU19J1zhNgK6cc6m3K6',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/bbc-indonesia/id434812049',
-        linkType: 'apple',
-      },
-    ],
-    p08wdlfl: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/1d4NQ20PkCliBnJn4NHeHO',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/global-newsbeat/id1537379741',
-        linkType: 'apple',
-      },
-    ],
-    p0btnmzx: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/4bSkrXP20pPmiF4Mjb2Wuw',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/investigasi-skandal-adopsi/id1613760117',
-        linkType: 'apple',
-      },
-    ],
-    p0dbsl80: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/43zcFT0Ac70PogTlN8v8jU',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/flora-carita/id1652325843',
-        linkType: 'apple',
-      },
-    ],
-  },
-};
-
-export default externalLinks;
