--- src/app/routes/onDemandAudio/tempData/podcastExternalLinks/nepali.js
@@ -1,19 +1,0 @@
-const externalLinks = {
-  default: {
-    p02pc9w3: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/2oE8PaJfhvymftvgAxHP2d',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/%E0%A4%AC-%E0%A4%AC-%E0%A4%B8-%E0%A4%A8-%E0%A4%AA-%E0%A4%B2-%E0%A4%AA%E0%A4%A1%E0%A4%95-%E0%A4%B8-%E0%A4%9F/id475748640',
-        linkType: 'apple',
-      },
-    ],
-  },
-};
-
-export default externalLinks;
