--- src/app/routes/onDemandAudio/tempData/podcastExternalLinks/russian.js
@@ -1,126 +1,0 @@
-const externalLinks = {
-  default: {
-    p05607v8: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/4J0FwkJYlSvkqZrJWdG9nt',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%D1%80%D0%B0%D0%B4%D0%B8%D0%BE%D0%B0%D1%80%D1%85%D0%B8%D0%B2-%D1%80%D1%83%D1%81%D1%81%D0%BA%D0%BE%D0%B9-%D1%81%D0%BB%D1%83%D0%B6%D0%B1%D1%8B-%D0%B1%D0%B8-%D0%B1%D0%B8-%D1%81%D0%B8/id1248943501',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Castbox',
-        linkUrl:
-          'https://castbox.fm/channel/%D0%A0%D0%B0%D0%B4%D0%B8%D0%BE%D0%B0%D1%80%D1%85%D0%B8%D0%B2-%D0%A0%D1%83%D1%81%D1%81%D0%BA%D0%BE%D0%B9-%D1%81%D0%BB%D1%83%D0%B6%D0%B1%D1%8B-%D0%91%D0%B8-%D0%B1%D0%B8-%D1%81%D0%B8-id933314?country=us',
-        linkType: 'castbox',
-      },
-    ],
-    p076qqzl: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/2CR7Wn6IzBVXPxQ5uzkkz1',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/ru/podcast/%D1%87%D1%82%D0%BE-%D1%8D%D1%82%D0%BE-%D0%B1%D1%8B%D0%BB%D0%BE/id1460240829?l=en',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Castbox',
-        linkUrl: 'https://castbox.fm/vc/2092482',
-        linkType: 'castbox',
-      },
-    ],
-    p08pxjzf: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/0g5kVHIpOM8V4qtgCqFppm',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%D0%B0%D0%BC%D0%B5%D1%80%D0%B8%D0%BA%D0%B0%D0%BD%D1%81%D0%BA%D0%B8%D0%B9-%D0%BF%D0%B8%D1%80%D0%BE%D0%B3/id1529622434',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Castbox',
-        linkUrl: 'https://castbox.fm/vc/3283439',
-        linkType: 'castbox',
-      },
-    ],
-    p075ydnq: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/3vu4FSOf79BHsVQ5sV0LmV',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/us/podcast/%D1%80%D0%B0%D0%B7%D0%B3%D0%BE%D0%B2%D0%BE%D1%80%D1%8B-%D1%81-%D0%B0%D1%80%D0%B1%D0%B8%D1%82%D1%80%D0%BE%D0%BC/id1459245691',
-        linkType: 'apple',
-      },
-    ],
-    p08l9l66: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/1mflnDxXdhrxp2BKC782Im',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/gb/podcast/%D0%B1%D0%B8-%D0%B1%D0%B8-%D1%81%D0%B8-doc/id1524273697',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Castbox',
-        linkUrl:
-          'https://castbox.fm/channel/id3130703?utm_campaign=i_share_ch&utm_medium=dlink&utm_source=i_share&country=us',
-        linkType: 'castbox',
-      },
-    ],
-    p09nx6zm: [
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/ru/podcast/%D0%BF%D0%B0%D1%81%D1%81%D0%B0%D0%B6%D0%B8%D1%80%D1%8B-%D0%B9%D0%BE%D0%BC%D0%B5%D0%B9-%D0%BC%D0%B0%D1%80%D1%83/id1575896140',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Castbox',
-        linkUrl:
-          'https://castbox.fm/channel/%D0%9F%D0%B0%D1%81%D1%81%D0%B0%D0%B6%D0%B8%D1%80%D1%8B-%D0%99%D0%BE%D0%BC%D0%B5%D0%B9-%D0%9C%D0%B0%D1%80%D1%83-id4407848',
-        linkType: 'castbox',
-      },
-    ],
-    p082h9l8: [
-      {
-        linkText: 'Spotify',
-        linkUrl: 'https://open.spotify.com/show/0gNOG2xiadfkRrBJnubfHT',
-        linkType: 'spotify',
-      },
-      {
-        linkText: 'Apple',
-        linkUrl:
-          'https://podcasts.apple.com/ru/podcast/8-%D0%B8%D1%81%D1%82%D0%BE%D1%80%D0%B8%D0%B9-%D0%B8%D0%B7-90-%D1%85/id1470221669',
-        linkType: 'apple',
-      },
-      {
-        linkText: 'Castbox',
-        linkUrl:
-          'https://castbox.fm/channel/\u041a\u043e\u043c\u0430\u043d\u0434\u0438\u0440\u043e\u0432\u043a\u0430-\u0432-\u0414\u043e\u043d\u0435\u0446\u043a-id2616795',
-        linkType: 'castbox',
-      },
-    ],
-  },
-};
-
-export default externalLinks;
